Refactoring

+++ internal/dto/image.go
@@ -1,11 +1,11 @@
 package dto
 
 type ImageResponse struct {
-	ID        int    `json:"id"`
-	Image     string `json:"image"`
-	PlaceID   *int   `json:"place_id"`
-	ThingID   *int   `json:"thing_id"`
-	CreatedAt string `json:"created_at"`
+	ID        uint64  `json:"id"`
+	Image     string  `json:"image"`
+	PlaceID   *uint64 `json:"place_id"`
+	ThingID   *uint64 `json:"thing_id"`
+	CreatedAt string  `json:"created_at"`
 }
 
 type ImagesResponse struct {
