+++ internal/mappers/thing_image.go
@@ -2,7 +2,7 @@ package mappers
 
 import "git.dmitriygnatenko.ru/dima/homethings/internal/models"
 
-func ToAddThingImageRequest(thingID int, image string) models.AddThingImageRequest {
+func ToAddThingImageRequest(thingID uint64, image string) models.AddThingImageRequest {
 	return models.AddThingImageRequest{
 		ThingID: thingID,
 		Image:   image,
